+++ CHANGES.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+* Improve audio on android by using `AsynchronousAndroidAudio`
+
 ### Removed
 
 ---